backend/requirements-test.txt
# Testing dependencies
pytest==8.3.4
pytest-flask==1.3.0
pytest-cov==6.0.0
pytest-mock==3.14.0
responses==0.25.3  # For mocking HTTP requests

# Linting
ruff>=0.9.0